XC-293: Add Rustdoc for all new types
