+++ libs/client/src/fixtures/mod.rs
@@ -4,8 +4,7 @@
 
 use crate::{ClientBuilder, Runtime};
 use async_trait::async_trait;
-use candid::utils::ArgumentEncoder;
-use candid::{CandidType, Principal};
+use candid::{utils::ArgumentEncoder, CandidType, Principal};
 use ic_cdk::api::call::RejectionCode;
 use serde::de::DeserializeOwned;
 use sol_rpc_types::{AccountData, AccountEncoding, AccountInfo};
