+++ libs/types/src/solana/request/mod.rs
@@ -167,12 +167,22 @@ impl GetSlotParams {
     }
 }
 
+/// The parameters for a Solana [`getTransaction`](https://solana.com/docs/rpc/http/gettransaction) RPC method call.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType)]
 pub struct GetTransactionParams {
+    /// Transaction signature.
     pub signature: Signature,
+    /// Specifies the required finality of the transaction being queried.
     pub commitment: Option<CommitmentLevel>,
+    /// Set the max transaction version to return in responses.
+    ///
+    /// If the requested transaction is a higher version, an error will be returned. If this
+    /// parameter is omitted, only legacy transactions will be returned, and any versioned
+    /// transaction will prompt the error.
     #[serde(rename = "maxSupportedTransactionVersion")]
     pub max_supported_transaction_version: Option<u8>,
+    /// Encoding for the returned transaction
+    // TODO XC-293: Add notes about `jsonParsed` from https://solana.com/de/docs/rpc/http/gettransaction
     pub encoding: Option<GetTransactionEncoding>,
 }
 
@@ -200,11 +210,15 @@ impl From<solana_signature::Signature> for GetTransactionParams {
     }
 }
 
+/// Encoding format for the returned transaction from a [`getTransaction`](https://solana.com/docs/rpc/http/gettransaction)`
+/// RPC method call.
 // TODO XC-343: Add support for `json` and `jsonParsed` formats.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType)]
 pub enum GetTransactionEncoding {
+    /// The transaction is base64-encoded.
     #[serde(rename = "base64")]
     Base64,
+    /// The transaction is base58-encoded.
     #[serde(rename = "base58")]
     Base58,
 }
