+++ libs/types/src/solana/transaction/instruction.rs
@@ -5,9 +5,14 @@ use solana_transaction_status_client_types::{
     UiCompiledInstruction, UiInnerInstructions, UiInstruction,
 };
 
+/// List of [inner instructions](https://solana.com/de/docs/rpc/json-structures#inner-instructions)
+/// for a Solana transaction.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub struct InnerInstructions {
+    ///  Index of the transaction instruction from which the inner instruction(s) originated.
     pub index: u8,
+    /// Ordered list of inner program instructions that were invoked during a single transaction
+    /// instruction.
     pub instructions: Vec<Instruction>,
 }
 
@@ -26,9 +31,32 @@ impl TryFrom<UiInnerInstructions> for InnerInstructions {
     }
 }
 
+impl From<InnerInstructions> for UiInnerInstructions {
+    fn from(instructions: InnerInstructions) -> Self {
+        Self {
+            index: instructions.index,
+            instructions: instructions
+                .instructions
+                .into_iter()
+                .map(Into::into)
+                .collect(),
+        }
+    }
+}
+
+/// A directive for a single invocation of a Solana program.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub enum Instruction {
+    /// A compiled Solana instruction.
     Compiled(CompiledInstruction),
+}
+
+impl From<Instruction> for UiInstruction {
+    fn from(instruction: Instruction) -> Self {
+        match instruction {
+            Instruction::Compiled(compiled) => Self::Compiled(compiled.into()),
+        }
+    }
 }
 
 impl TryFrom<UiInstruction> for Instruction {
@@ -44,18 +72,39 @@ impl TryFrom<UiInstruction> for Instruction {
     }
 }
 
+/// Represents a compiled [instruction](https://solana.com/de/docs/references/terminology#instruction)
+/// as part of a Solana transaction.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub struct CompiledInstruction {
+    /// Index into the transaction `message.accountKeys` array (see the transaction
+    /// [JSON structure](https://solana.com/de/docs/rpc/json-structures#transactions)) indicating
+    /// the program account that executes this instruction.
     #[serde(rename = "programIdIndex")]
     pub program_id_index: u8,
+    /// List of ordered indices into the transaction `message.accountKeys` array (see the transaction
+    /// [JSON structure](https://solana.com/de/docs/rpc/json-structures#transactions)) indicating
+    /// which accounts to pass to the program.
     pub accounts: Vec<u8>,
+    /// The program input data encoded in a base-58 string.
     pub data: String,
+    /// The stack height at which this instruction was invoked during cross-program execution.
     #[serde(rename = "stackHeight")]
     pub stack_height: Option<u32>,
 }
 
 impl From<UiCompiledInstruction> for CompiledInstruction {
     fn from(instruction: UiCompiledInstruction) -> Self {
+        Self {
+            program_id_index: instruction.program_id_index,
+            accounts: instruction.accounts,
+            data: instruction.data,
+            stack_height: instruction.stack_height,
+        }
+    }
+}
+
+impl From<CompiledInstruction> for UiCompiledInstruction {
+    fn from(instruction: CompiledInstruction) -> Self {
         Self {
             program_id_index: instruction.program_id_index,
             accounts: instruction.accounts,
