+++ libs/types/src/solana/transaction/mod.rs
@@ -5,23 +5,33 @@ pub mod reward;
 use crate::{Pubkey, RpcError, Slot, Timestamp};
 use candid::{CandidType, Deserialize};
 use error::TransactionError;
-use instruction::{CompiledInstruction, InnerInstructions, Instruction};
+use instruction::InnerInstructions;
 use reward::Reward;
 use serde::Serialize;
 use solana_account_decoder_client_types::token::UiTokenAmount;
 use solana_transaction_status_client_types::{
     option_serializer::OptionSerializer, EncodedConfirmedTransactionWithStatusMeta,
-    EncodedTransactionWithStatusMeta, UiCompiledInstruction, UiInnerInstructions, UiInstruction,
-    UiReturnDataEncoding, UiTransactionReturnData, UiTransactionStatusMeta,
+    EncodedTransactionWithStatusMeta, UiReturnDataEncoding, UiTransactionReturnData,
+    UiTransactionStatusMeta,
 };
 
+/// Solana [transaction](https://solana.com/docs/references/terminology#transaction) information
+/// as returned by the [`getTransaction`](https://solana.com/de/docs/rpc/http/gettransaction) RPC
+/// method.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub struct TransactionInfo {
+    /// The slot this transaction was processed in.
     pub slot: Slot,
+    /// Estimated production time of when the transaction was processed. [`None`] if not available
     #[serde(rename = "blockTime")]
     pub block_time: Option<Timestamp>,
+    /// Transaction status [metadata](https://solana.com/de/docs/rpc/json-structures#transaction-status-metadata)
+    /// object or [`None`].
     pub meta: Option<TransactionStatusMeta>,
+    /// [Transaction](https://solana.com/de/docs/rpc/json-structures#transactions) object, either
+    /// in JSON format or encoded binary data, depending on encoding parameter.
     pub transaction: EncodedTransaction,
+    /// Transaction version. [`None`] if [`maxSupportedTransactionVersion`] is not set in request params.
     pub version: Option<TransactionVersion>,
 }
 
@@ -59,27 +69,47 @@ impl From<TransactionInfo> for EncodedConfirmedTransactionWithStatusMeta {
     }
 }
 
+/// Transaction status [metadata](https://solana.com/de/docs/rpc/json-structures#transaction-status-metadata) object.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub struct TransactionStatusMeta {
+    /// [`Err`] if transaction failed, [`Ok`] if transaction succeeded.
     pub status: Result<(), TransactionError>,
+    /// Fee this transaction was charged.
     pub fee: u64,
+    /// Array of account balances from before the transaction was processed.
     #[serde(rename = "preBalances")]
     pub pre_balances: Vec<u64>,
+    /// Array of account balances after the transaction was processed.
     #[serde(rename = "postBalances")]
     pub post_balances: Vec<u64>,
+    /// List of [inner instructions](https://solana.com/de/docs/rpc/json-structures#inner-instructions)
+    /// or [`None`] if inner instruction recording was not enabled during this transaction.
     #[serde(rename = "innerInstructions")]
     pub inner_instructions: Option<Vec<InnerInstructions>>,
+    /// Array of log messages or [`None`] if log message recording was not enabled during this
+    /// transaction.
     #[serde(rename = "logMessages")]
     pub log_messages: Option<Vec<String>>,
+    /// List of [token balances](https://solana.com/de/docs/rpc/json-structures#token-balances) from
+    /// before the transaction was processed or [`None`] if token balance recording was not yet
+    /// enabled during this transaction.
     #[serde(rename = "preTokenBalances")]
     pub pre_token_balances: Option<Vec<TransactionTokenBalance>>,
+    /// List of [token balances](https://solana.com/de/docs/rpc/json-structures#token-balances) from
+    /// after the transaction was processed or [`None`] if token balance recording was not yet
+    /// enabled during this transaction.
     #[serde(rename = "postTokenBalances")]
     pub post_token_balances: Option<Vec<TransactionTokenBalance>>,
+    /// Array of transaction-level rewards.
     pub rewards: Option<Vec<Reward>>,
+    /// Transaction addresses loaded from address lookup tables. Undefined if `maxSupportedTransactionVersion`
+    /// is not set in request params, or if `jsonParsed` encoding is set in request params.
     #[serde(rename = "loadedAddresses")]
     pub loaded_addresses: Option<LoadedAddresses>,
+    /// The most-recent return data generated by an instruction in the transaction.
     #[serde(rename = "returnData")]
     pub return_data: Option<TransactionReturnData>,
+    /// Number of compute units consumed by the transaction
     #[serde(rename = "computeUnitsConsumed")]
     pub compute_units_consumed: Option<u64>,
 }
@@ -157,11 +187,15 @@ impl TryFrom<UiTransactionStatusMeta> for TransactionStatusMeta {
     }
 }
 
+/// [Transaction](https://solana.com/de/docs/rpc/json-structures#transactions) object, either in
+/// JSON format or encoded binary data.
 // TODO XC-343: Add variants corresponding to `Json` and `Accounts` in
 //  `solana_transaction_status_client_types::EncodedTransaction`.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub enum EncodedTransaction {
+    /// Legacy format kept for backwards compatibility. The transaction is base58-encoded.
     LegacyBinary(String),
+    ///The transaction is encoded in one of the [`TransactionBinaryEncoding`] formats.
     Binary(String, TransactionBinaryEncoding),
 }
 
@@ -191,10 +225,13 @@ impl From<EncodedTransaction> for solana_transaction_status_client_types::Encode
     }
 }
 
+/// Binary encoding format for an [`EncodedTransaction`].
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub enum TransactionBinaryEncoding {
+    /// The transaction is base64-encoded.
     #[serde(rename = "base64")]
     Base64,
+    /// The transaction is base58-encoded.
     #[serde(rename = "base58")]
     Base58,
 }
@@ -222,46 +259,20 @@ impl From<TransactionBinaryEncoding>
     }
 }
 
-impl From<InnerInstructions> for UiInnerInstructions {
-    fn from(instructions: InnerInstructions) -> Self {
-        Self {
-            index: instructions.index,
-            instructions: instructions
-                .instructions
-                .into_iter()
-                .map(Into::into)
-                .collect(),
-        }
-    }
-}
-
-impl From<Instruction> for UiInstruction {
-    fn from(instruction: Instruction) -> Self {
-        match instruction {
-            Instruction::Compiled(compiled) => Self::Compiled(compiled.into()),
-        }
-    }
-}
-
-impl From<CompiledInstruction> for UiCompiledInstruction {
-    fn from(instruction: CompiledInstruction) -> Self {
-        Self {
-            program_id_index: instruction.program_id_index,
-            accounts: instruction.accounts,
-            data: instruction.data,
-            stack_height: instruction.stack_height,
-        }
-    }
-}
-
+/// Represents the balance of a specific SPL token account.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub struct TransactionTokenBalance {
+    /// Index of the account in which the token balance is provided for.
     #[serde(rename = "accountIndex")]
     pub account_index: u8,
+    /// Pubkey of the token's mint.
     pub mint: String,
+    /// A human-readable representation of the token amount.
     #[serde(rename = "uiTokenAmount")]
     pub ui_token_amount: TokenAmount,
+    /// Pubkey of token balance's owner.
     pub owner: Option<Pubkey>,
+    /// Pubkey of the Token program that owns the account.
     #[serde(rename = "programId")]
     pub program_id: Option<Pubkey>,
 }
@@ -294,12 +305,17 @@ impl From<TransactionTokenBalance>
     }
 }
 
+/// A human-readable representation of a token amount.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub struct TokenAmount {
+    /// DEPRECATED: Token amount as a float, accounting for decimals.
     #[serde(rename = "uiAmount")]
     pub ui_amount: Option<f64>,
+    /// Number of decimals configured for token's mint.
     pub decimals: u8,
+    /// Raw amount of tokens as a string, ignoring decimals.
     pub amount: String,
+    /// Token amount as a string, accounting for decimals.
     #[serde(rename = "uiAmountString")]
     pub ui_amount_string: String,
 }
@@ -326,9 +342,12 @@ impl From<UiTokenAmount> for TokenAmount {
     }
 }
 
+/// Transaction addresses loaded from address lookup tables.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub struct LoadedAddresses {
+    /// Ordered list of base-58 encoded addresses for writable loaded accounts.
     pub writable: Vec<Pubkey>,
+    /// Ordered list of base-58 encoded addresses for readonly loaded accounts.
     pub readonly: Vec<Pubkey>,
 }
 
@@ -350,10 +369,13 @@ impl From<LoadedAddresses> for solana_transaction_status_client_types::UiLoadedA
     }
 }
 
+/// Represents the return data emitted by a program during transaction execution.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub struct TransactionReturnData {
+    /// The program that generated the return data.
     #[serde(rename = "programId")]
     pub program_id: Pubkey,
+    /// The return data itself, as base-64 encoded binary data.
     pub data: String,
 }
 
@@ -378,14 +400,12 @@ impl From<TransactionReturnData> for UiTransactionReturnData {
     }
 }
 
-#[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
-pub enum ReturnDataEncoding {
-    Base64,
-}
-
+/// Enum representing the version of a Solana transaction.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub enum TransactionVersion {
+    /// Legacy transaction format, which does not explicitly include a version number.
     Legacy,
+    /// Versioned transaction format.
     Number(u8),
 }
