+++ libs/types/src/solana/transaction/reward.rs
@@ -2,14 +2,21 @@ use crate::Pubkey;
 use candid::{CandidType, Deserialize};
 use serde::Serialize;
 
+/// Represents a reward or penalty applied to an account for fees, rent, voting, or staking activity.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub struct Reward {
+    /// The public key, of the account that received the reward.
     pub pubkey: Pubkey,
+    /// Number of reward lamports credited or debited by the account
     pub lamports: i64,
+    /// Account balance in lamports after the reward was applied.
     #[serde(rename = "postBalance")]
     pub post_balance: u64,
+    /// Type of reward.
     #[serde(rename = "rewardType")]
     pub reward_type: Option<RewardType>,
+    /// Vote account commission when the reward was credited, only present for voting and staking
+    /// rewards.
     pub commission: Option<u8>,
 }
 
@@ -37,11 +44,16 @@ impl From<Reward> for solana_transaction_status_client_types::Reward {
     }
 }
 
+/// Enum representing the type of reward granted to an account on the Solana network.
 #[derive(Debug, Clone, Deserialize, Serialize, CandidType, PartialEq)]
 pub enum RewardType {
+    /// Reward from transaction fees collected in the block.
     Fee,
+    /// Reward from rent fees paid by accounts storing data on-chain.
     Rent,
+    /// Reward earned from delegating stake to validators with good performance.
     Staking,
+    /// Reward earned for participating in vote transactions to help reach consensus.
     Voting,
 }
